Move valiationGroup to cache
