+++ web/src/app/modules/forms/modules/validation/services/validation.service.ts
@@ -14,9 +14,9 @@ import { ValidNameValidator } from '../../../../../validation/valid-name-validat
 import { ValidationErrorSeverity } from '../../../../../validation/validation-error-severity';
 import { ValidationResult } from '../../../../../validation/validation-result';
 import { NavigatorService } from '../../../../navigation/modules/navigator/services/navigator.service';
-import { ValidationCache, ValidationPair } from '../util/validation-cache';
+import { ValidationCache, ValidationGroup, ValidationPair } from '../util/validation-cache';
 
-export type ValidationGroup = { element: IContainer, messages: string[] };
+
 
 @Injectable()
 export class ValidationService extends Monitorable {
@@ -36,18 +36,9 @@ export class ValidationService extends Monitorable {
     public getValidationResults(parentElement: IContainer): ValidationPair[] {
         return parentElement ? this.validationCache.getValidationResults(parentElement.url) : [];
     }
+
     public getValidationResultsGrouped(parentElement: IContainer): ValidationGroup[] {
-        let validationResult = this.getValidationResults(parentElement);
-        let result: ValidationGroup[] = [];
-        for (const pair of validationResult) {
-            let resultItem = result.find(element => element.element === pair.element);
-            if (resultItem !== undefined) {
-                resultItem.messages.push(pair.message);
-            } else {
-                result.push({ element: pair.element, messages: [pair.message] });
-            }
-        }
-        return result;
+        return parentElement ? this.validationCache.getValidationResultsGrouped(parentElement.url) : [];
     }
 
     private static requiredFieldValidatorMap: { [className: string]: RequiredFieldsValidator };
